requirements.txt
pillow
tqdm
tkinter